requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
azure-search-documents==11.4.0
semantic-kernel==0.9.0
openai==1.3.0
azure-identity==1.15.0
pydantic==2.5.0
pydantic-settings==2.1.0
langdetect==1.0.9
python-multipart==0.0.6
httpx==0.25.2
tenacity==8.2.3
# Local vector database and embeddings
chromadb==0.4.22
sentence-transformers==2.2.2
# PDF processing
pypdf==3.17.4
pdfplumber==0.10.3
# Text processing
numpy==1.24.3
torch==2.1.0
